Create the Catalyst Cloud API client library (#1)
This is a new package which can be used to automatically install all required packages for interacting with Catalyst Cloud via the CLI.

Some basic package documentation has also been provided via `README.md` and `CHANGELOG.md`, which will be used to provide the long description on PyPI.

A GitHub Actions pipeline has been provided to automatically publish releases to PyPI and GitHub. Just tag the version (e.g. `1.0.0`) in the git repository, and the pipeline will automatically kick in.